# doc/swagger/account-request.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example NestJS App",
        "description": "Example NestJS application, with several modules",
        "version": "0.0.1",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
...
            "CountryBasicData": {
                "type": "object",
                "properties": {
                    "countryCode": {
                        "type": "string",
                        "description": "ISO-3 code of the country",
                        "example": "ARG"
                    },
                    "countryName": {
                        "type": "string",
                        "description": "Country name in Spanish",
                        "example": "Argentina"
                    },
                    "population": {
                        "type": "number",
                        "description": "Country population as of recent estimate",
                        "example": 45200000
                    }
                },
                "required": [
                    "countryCode",
                    "countryName",
                    "population"
                ]
            },
            "CountryLongSummary": {
                "type": "object",
                "properties": {
                    "countryCode": {
                        "type": "string",
                        "description": "ISO-3 code of the country",
                        "example": "ARG"
                    },
                    "countryName": {
                        "type": "string",
                        "description": "Country name in Spanish",
                        "example": "Argentina"
                    },
                    "population": {
                        "type": "number",
                        "description": "Country population as of recent estimate",
                        "example": 45200000
                    },
                    "internetDomain": {
                        "type": "string",
                        "description": "Main internet domain assigned to this country",
                        "example": ".ar"
                    },
                    "currency": {
                        "type": "object",
                        "description": "Data about the main currency of the country",
                        "properties": {
                            "code": {
                                "description": "Currency ISO code",
                                "example": "ARS",
                                "type": "string"
                            },
                            "name": {
                                "description": "Currency name",
                                "example": "Peso argentino",
                                "type": "string"
                            },
                            "byUSD": {
                                "description": "Rate in units for one US Dollar",
                                "example": 82.5,
                                "type": "number"
                            }
                        }
                    },
                    "neighbors": {
                        "description": "Data about the neighbor countries",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CountryBasicData"
                        }
                    },
                    "totalNeighborPopulation": {
                        "type": "number",
                        "description": "Sum of the population of the neighbors has this country",
                        "example": 280000000
                    },
                    "covid19LastRecord": {
                        "type": "object",
                        "description": "Last record of COVID-related information",
                        "properties": {
                            "confirmed": {
                                "description": "Total confirmed cases",
                                "example": 200000,
                                "type": "number"
                            },
                            "active": {
                                "description": "Currently active cases",
                                "example": 45000,
                                "type": "number"
                            },
                            "recovered": {
                                "description": "Persons that have been recovered from Covid",
                                "example": 152000,
                                "type": "number"
                            },
                            "deaths": {
                                "description": "Persons dead of Covid",
                                "example": 3000,
                                "type": "number"
                            },
                            "date": {
                                "description": "Covid report date",
                                "example": "2020-05-28",
                                "type": "string",
                                "format": "YYYY-MM-DD"
                            }
                        }
                    }
                },
                "required": [
                    "countryCode",
                    "countryName",
                    "population",
                    "internetDomain",
                    "currency",
                    "neighbors",
                    "totalNeighborPopulation"
                ]
            },
            "CountryShortSummary": {
                "type": "object",
                "properties": {
                    "countryCode": {
                        "type": "string",
                        "description": "ISO-3 code of the country",
                        "example": "ARG"
                    },
                    "countryName": {
                        "type": "string",
                        "description": "Country name in Spanish",
                        "example": "Argentina"
                    },
                    "population": {
                        "type": "number",
                        "description": "Country population as of recent estimate",
                        "example": 45200000
                    },
                    "internetDomain": {
                        "type": "string",
                        "description": "Main internet domain assigned to this country",
                        "example": ".ar"
                    },
                    "currencyCode": {
                        "type": "string",
                        "description": "ISO code of the main currency of the country",
                        "example": "ARS"
                    },
                    "neighborCountryCount": {
                        "type": "number",
                        "description": "How many neighbors has this country",
                        "example": 5
                    }
                },
                "required": [
                    "countryCode",
                    "countryName",
                    "population",
                    "internetDomain",
                    "currencyCode",
                    "neighborCountryCount"
                ]
            },
...
        }
    },
    "paths": {
...
        "/countries/{countryCode}": {
            "get": {
                "operationId": "CountryDataController_getCountryData",
                "summary": "",
                "description": "Get data about a specific country, representation can be chosen among several options",
                "parameters": [
                    {
                        "name": "countryCode",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data delivered",
                        "content": {
                            "application/vnd.bdsol.countryShortSummary+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CountryShortSummary"
                                }
                            },
                            "application/vnd.bdsol.countryLongSummary+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CountryLongSummary"
                                }
                            },
                            "application/vnd.bdsol.countryTextDescription+json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Countries"
                ]
            }
        },
...
        "/countries/{countryCode}/shortSummary": {
            "get": {
                "operationId": "CountryDataController_getShortSummaryEndpoint",
                "summary": "",
                "description": "Get limited data about a specific country",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "header",
                        "description": "Id of the user who makes the request",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data delivered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CountryShortSummary"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "Info about a very bad country is requested"
                    }
                },
                "tags": [
                    "Countries"
                ]
            }
        },
        "/countries/many/{countryCodes}/shortSummary": {
            "get": {
                "operationId": "CountryDataController_getManyCountriesShortSummary",
                "parameters": [
                    {
                        "name": "countryCodes",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Countries"
                ]
            }
        },
        "/countries/{countryCode}/info": {
            "get": {
                "operationId": "CountryDataController_getInfo",
                "parameters": [
                    {
                        "name": "countryCode",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Countries"
                ]
            }
        },
...
        "/countries/{countryCode}/infoWithCovid": {
            "get": {
                "operationId": "CountryDataController_getInfoWithCovid",
                "parameters": [
                    {
                        "name": "countryCode",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Countries",
                    "Countries + Covid"
                ]
            }
        },
...
        "/countries/{countryCode}/covid": {
            "get": {
                "operationId": "CountryDataController_getCovidData",
                "parameters": [
                    {
                        "name": "countryCode",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fromDate",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "toDate",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Countries",
                    "Countries + Covid"
                ]
            }
        },
...
        "/agencies/byText": {
            "get": {
                "operationId": "AgencyController_getAgencyByTextSearch",
                "parameters": [
                    {
                        "name": "q",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
...
        "/agencies/{code}": {
            "get": {
                "operationId": "AgencyController_getAgencyByCode",
                "parameters": [
                    {
                        "name": "code",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/agencies/{code}/range": {
            "get": {
                "operationId": "AgencyController_getAgencyByCodeRange",
                "parameters": [
                    {
                        "name": "code",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/agencies/massiveAddition": {
            "post": {
                "operationId": "AgencyController_agencyMassiveAddition",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        }
    }
}